increase font size of description in hero Fixes #64
