+++ src/payload-types.ts
@@ -183,12 +183,7 @@ export interface Page {
  */
 export interface HeroBlockType {
   image: number | Media;
-  stories: {
-    story?: string | null;
-    id?: string | null;
-  }[];
   title?: string | null;
-  description?: string | null;
   id?: string | null;
   blockName?: string | null;
   blockType: 'heroBlock';
@@ -434,14 +429,7 @@ export interface PagesSelect<T extends boolean = true> {
  */
 export interface HeroBlockTypeSelect<T extends boolean = true> {
   image?: T;
-  stories?:
-    | T
-    | {
-        story?: T;
-        id?: T;
-      };
   title?: T;
-  description?: T;
   id?: T;
   blockName?: T;
 }
